Crude HELO-time protocol-version adoption.
Assume all the protocol "versions" follow the same HELO argument pattern
( chr(version)+b'\0\0\1', nil, nil, nil ) and always assume whatever
version response we get back from the device in the first
interaction is the correct one, thenceforth, apeing it by default.

This way, we can take this as a "hint" as to how to interact with the
device (presumably, that's not abuse, but the whole point of it).

Furthermore, if the protocol version is 4 or greater, do not attempt
to funnel interactive commands through 'sh' (the busybox-alike shell)
but pass them through unadorned to 'EXEC' (otherwise they are all
rejected by lafd due to its strict $0 filter for EXEC commands).

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>